dec and undec discrete wavelet (framelet) transforms now in Matlab and in c.
The basic filterbank (to be iterated) can contain an arbitrary number of filters (channels) each of which can be followed by an arbitrary (but equal) subsampling factor (and vice versa for the synthesis).

The filters followed by non-equal subsampling factors (so far just wfilt_hden.m, three filters, factors 2,2,1) implemented just in Matlab. In this case, it seems the input signal have to be padded for the 'per' (non-expansive) transform, otherwise the periodic extension differs for each sub. fac.