Filterbank(real)dual/tight are fixed
The output filterbank format is now a struct with .h or .H fields.
It is possible to create a dual filterbank for some L and use it for
different L regardless of the filter type (assuming you do not expect a
perfect rec).
As soon as g.H becomes numeric it is also expected that there is another
field g.L, which stores L which was used to evaluate g.H(L).