Latest update, comments below:
For *ptpfun*:

Multiplicities of weights *w* were not dealt with correctly. We fixed
this and added the necessary code.

We added both cases of having only positive weights or only negative
weights.

We corrected the leading comment lines regarding the decay properties of
the continuous TP functions.

The current version has still some problems with roundoff errors if
negative weights with multiplicities are given, such as w=[-1,-1,-1]. We
will try to circumvent this problem.

For *ptpfundual*:

The multiplicities of weights were already correct.

We added both cases of having only positive weights or only negative
weights.

We ran several numerical tests with the following results: Our
computation of the dual is very accurate. Even if the *matchscale*
option gives a negative result (Assertion failed …..), our own test of
the Wexler-Raz identity shows that g and gamma are dual to each other.
We include our script *Wrtest* for you to check. This is
a very simple version, not suitable for large *L*. Therefore, we
recommend that the matchscale-option be deactivated in *ptpfundual*.

For the same reason, we believe that the ltfat program *gabdualnorm*
does not work accurately enough for our duals.

For both programs, we preferred that the *ptpfun* window has our
normalization.

Then, the computed  *gamma*  is a dual, without any other normalization.

Moreover, we changed the scripts:

*myknt2mlt*     into    *wmult*:    copyright of the original knt2mlt is
with C. de Boor, It is part of the spline-toolbox of matlab.

*divdiff_vector*:   we do not call   *wmult*, but provide the
multiplicities as a new input parameter